move rails cli actions to their own file move rails dependencies to their own file
