+++ lib/trimtool/cli/BootstrapGemfile
