+++ lib/trimtool/cli/Dockerfile
