+++ lib/trimtool/cli/docker-compose.yml
